+++ projekt/src/main/java/ba/unsa/etf/si/projekt/entiteti/Karta.java
@@ -14,16 +14,16 @@ public class Karta implements java.io.Serializable{
 	public void setOdrediste(String odrediste) {
 		this.odrediste = odrediste;
 	}
-	public double getVrijemePolaska_sati() {
+	public int getVrijemePolaska_sati() {
 		return vrijemePolaska_sati;
 	}
-	public void setVrijemePolaska_sati(double vrijemePolaska_sati) {
+	public void setVrijemePolaska_sati(int vrijemePolaska_sati) {
 		this.vrijemePolaska_sati = vrijemePolaska_sati;
 	}
-	public double getVrijemePolaska_minute() {
+	public int getVrijemePolaska_minute() {
 		return vrijemePolaska_minute;
 	}
-	public void setVrijemePolaska_minute(double vrijemePolaska_minute) {
+	public void setVrijemePolaska_minute(int vrijemePolaska_minute) {
 		this.vrijemePolaska_minute = vrijemePolaska_minute;
 	}
 	public TipKarte getTipKarte() {
@@ -57,8 +57,8 @@ public class Karta implements java.io.Serializable{
 		this.cijena = cijena;
 	}
 	String odrediste;
-	double vrijemePolaska_sati;
-	double vrijemePolaska_minute;
+	int vrijemePolaska_sati;
+	int vrijemePolaska_minute;
 	TipKarte tipKarte;
 	int datumPolaska_dan;
 	int datumPolaska_mjesec;
